requirements.txt
# Prometheus V3.5 Supreme - Complete Dependencies
# Generated: 2025-11-20
# Python 3.11+ required

# ============================================
# Core Dependencies
# ============================================
python-dotenv>=1.0.0
requests>=2.31.0
docker>=7.0.0
open-interpreter>=0.3.0

# ============================================
# AI Providers
# ============================================
anthropic>=0.18.0
openai>=1.12.0
litellm>=1.30.0

# ============================================
# Knowledge & Memory (ChromaDB)
# ============================================
chromadb>=0.4.22
sentence-transformers>=2.5.0

# ============================================
# Web & API
# ============================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0

# ============================================
# Browser Automation
# ============================================
playwright>=1.40.0

# ============================================
# Data Processing
# ============================================
pyyaml>=6.0.0
tiktoken>=0.6.0
shortuuid>=1.0.11
markdown-it-py>=3.0.0
regex>=2023.12.0

# ============================================
# Testing
# ============================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# ============================================
# UI (Desktop) - tkinter is built-in on Windows
# ============================================
# tkinter - Built-in on Windows Python

# ============================================
# Database - sqlite3 is built-in
# ============================================
# sqlite3 - Built-in Python

# ============================================
# Optional - Install only if needed
# ============================================
# spacy>=3.7.0  # For NLP Task Analyzer (optional)
# supabase>=2.0.0  # Legacy - system uses ChromaDB now
# redis>=5.0.0  # Optional cache (uses in-memory fallback)
